UserDetails

+++ EmployeeEvaluation/ClientApp/src/app/components/users/users-table/users-table.component.ts
@@ -1,5 +1,6 @@
 import { Component, Injectable, OnInit } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
+import { Guid } from 'guid-typescript';
 import { Project } from 'src/app/models/project.model';
 import { UserDTO } from 'src/app/models/users.model';
 import { ProjectsService } from 'src/app/services/projects.service';
@@ -11,7 +12,7 @@ import { UsersService } from 'src/app/services/users.service';
   styleUrls: ['./users-table.component.css'],
 })
 export class UsersTableComponent implements OnInit {
-  addUserFormGroup = new FormGroup({
+  editUserFormGroup = new FormGroup({
     nameControl: new FormControl(''),
     emailControl: new FormControl(''),
     roleControl: new FormControl(''),
@@ -19,6 +20,7 @@ export class UsersTableComponent implements OnInit {
   users: UserDTO[] = [];
   user!: UserDTO;
   projects: Project[] = [];
+  currentUserId!: Guid;
   constructor(
     private usersService: UsersService,
     private projectsService: ProjectsService
@@ -47,34 +49,40 @@ export class UsersTableComponent implements OnInit {
       },
     });
   }
-  httpGetUserById(id: string) {
-    this.usersService.getUserById(id).subscribe({
-      next: (user) => {
-        this.user = user;
-      },
-      error: (error) => {
-        console.log(error);
-      },
-    });
+  setCurrentUserId(id: Guid) {
+    this.currentUserId = id;
   }
-
-  httpDeleteUser(id: string) {
-    for (let i = 0; i < this.users.length; i++) {
-      if (this.users[i].id == id) this.users.splice(i, 1);
+  httpGetUserById(id?: Guid) {
+    if (id !== undefined) {
+      this.usersService.getUserById(id).subscribe({
+        next: (user) => {
+          this.user = user;
+        },
+        error: (error) => {
+          console.log(error);
+        },
+      });
     }
-    this.usersService.deleteUser(id).subscribe({
-      next: (response) => {
-        console.log(response);
-      },
-    });
   }
 
-  httpEditUser(id: string) {
-    this.httpGetUserById(id);
-    this.user.name = this.addUserFormGroup.controls.nameControl.value!;
-    this.user.email = this.addUserFormGroup.controls.emailControl.value!;
-    this.user.role = this.addUserFormGroup.controls.roleControl.value!;
-    this.usersService.editUser(id, this.user).subscribe({
+  httpDeleteUser(id?: Guid) {
+    if (id !== undefined) {
+      for (let i = 0; i < this.users.length; i++) {
+        if (this.users[i].id == id) this.users.splice(i, 1);
+      }
+      this.usersService.deleteUser(id).subscribe({
+        next: (response) => {
+          console.log(response);
+        },
+      });
+    }
+  }
+
+  httpEditUser() {
+    this.user.name = this.editUserFormGroup.controls.nameControl.value!;
+    this.user.email = this.editUserFormGroup.controls.emailControl.value!;
+    this.user.role = this.editUserFormGroup.controls.roleControl.value!;
+    this.usersService.editUser(this.user.id!, this.user).subscribe({
       next: (response) => {
         console.log(response);
       },
